added code to distinguish between OSC messages of different lengths and paths
The main branch accepts any OSC messages and tries to interpret it as having 3 floats. In addition, it only accepts from a specific IP address. This version requires a 3-float message. When the path starts with /l (presumably /light/something) it assumes it will be in the range 0-1.0 . Otherwise, the firmware calibrates by listening for the min and max values in each channel. This is pretty useful for--say--piping raw accelerometer data to the light via OSC. Resetting the light resets the adjustment, but you can still use any /l commands normally without reboot. I have also made it so any IP can send OSC to the light, which may be a 'security' and 'stability' risk but makes it much easier when just starting out. It is recommended to use a specific IP if you are on a network with other traffic.